Safari-609.1.11.1/LayoutTests/fast/url/user-visible/arabic_kasra.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("Test IDN behavior for Arabic Kasra letter.");

function test(url)
{
    var domURL = new URL(url);
    return internals.userVisibleString(domURL);
}

shouldBe("test('https://xn--apple-gkh.com/')", "'https://xn--apple-gkh.com/'");
shouldBe("test('http://xn--google-yri.com/')", "'http://xn--google-yri.com/'");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
